// LocalGovernmentUnit/src/types/gse.types.ts
export interface DisbursementVoucher {
  id: string;
  uuid: string;
  user_facility: string | null;
  pq: string | null;
  payee: string | null;
  desc: string;
  amount: number;
  dv_no: string;
  editable: boolean;
  status: boolean;
  mode_of_payment: string;
}